+++ src/finalisers/es6.js
@@ -10,10 +10,20 @@ export default function es6 ( bundle, magicString ) {
 			const specifiers = [];
 			const specifiersList = [specifiers];
 			const importedNames = keys( module.declarations )
-				.filter( name => name !== '*' && name !== 'default' );
+				.filter( name => name !== '*' && name !== 'default' )
+				.map( name => {
+					const declaration = module.declarations[ name ];
+
+					if ( declaration.name === declaration.safeName ) return declaration.name;
+					return `${declaration.name} as ${declaration.safeName}`;
+				});
 
 			if ( module.declarations.default ) {
-				specifiers.push( module.name );
+				if ( module.exportsNamespace ) {
+					specifiersList.push([ `${module.name}__default` ]);
+				} else {
+					specifiers.push( module.name );
+				}
 			}
 
 			const namespaceSpecifier = module.declarations['*'] ? `* as ${module.name}` : null;