+++ src/finalisers/shared/esModuleExport.js
@@ -1,0 +1,1 @@
+export default `Object.defineProperty(exports, '__esModule', { value: true });`;