+++ test/form/assignment-to-exports-class-declaration/_expected/amd.js
@@ -3,4 +3,6 @@ define(['exports'], function (exports) { 'use strict';
 	exports.Foo = class Foo {}
 	exports.Foo = lol( exports.Foo );
 
+	Object.defineProperty(exports, '__esModule', { value: true });
+
 });