+++ test/form/computed-properties/_expected/umd.js
@@ -20,4 +20,6 @@
 	exports.x = x;
 	exports.X = X;
 
+	Object.defineProperty(exports, '__esModule', { value: true });
+
 }));