+++ test/form/conflicting-imports/_expected/cjs.js
@@ -1,0 +1,8 @@
+'use strict';
+
+var foo = require('foo');
+var bar = require('bar');
+
+console.log( bar.a );
+
+console.log( foo.a );