+++ test/form/dedupes-external-imports/_expected/umd.js
@@ -33,4 +33,6 @@
 	exports.bar = bar;
 	exports.baz = baz;
 
+	Object.defineProperty(exports, '__esModule', { value: true });
+
 }));