+++ test/form/export-default-import/_expected/umd.js
@@ -10,4 +10,6 @@
 
 	exports.x = x;
 
+	Object.defineProperty(exports, '__esModule', { value: true });
+
 }));