+++ test/form/exports-at-end-if-possible/_expected/cjs.js
@@ -1,5 +1,7 @@
 'use strict';
 
+Object.defineProperty(exports, '__esModule', { value: true });
+
 var FOO = 'foo';
 
 console.log( FOO );