+++ test/form/import-external-namespace-and-default/_expected/cjs.js
@@ -1,0 +1,8 @@
+'use strict';
+
+var foo = require('foo');
+var foo__default = foo['default'];
+
+console.log( foo.bar );
+
+console.log( foo__default );