+++ test/form/no-treeshake/_expected/umd.js
@@ -26,4 +26,6 @@
 	exports.getPrototypeOf = getPrototypeOf;
 	exports.strange = quux;
 
+	Object.defineProperty(exports, '__esModule', { value: true });
+
 }));